TFLM: Fix double-promotion error.
Some of these double promotion is not obvious as va_args implicitly promotes float to double.

PiperOrigin-RevId: 290883516

+++ tensorflow/lite/experimental/ruy/check_macros.h
@@ -35,7 +35,7 @@ struct ToString {
 template <>
 struct ToString<float, void> {
   static void Run(float value, char* buf) {
-    snprintf(buf, kValueBufSize, "%.9g", static_cast<double>(value));
+    snprintf(buf, kValueBufSize, "%.9g", value);
   }
 };
