+++ tensorflow/lite/micro/examples/hello_world/output_handler.cc
@@ -18,7 +18,5 @@ limitations under the License.
 void HandleOutput(tflite::ErrorReporter* error_reporter, float x_value,
                   float y_value) {
   // Log the current X and Y values
-  error_reporter->Report("x_value: %f, y_value: %f\n",
-                         static_cast<double>(x_value),
-                         static_cast<double>(y_value));
+  error_reporter->Report("x_value: %f, y_value: %f\n", x_value, y_value);
 }
